chore: Move service files into individual folders
